moved LoginAction and LogoutAction to their own files.
The intent is to eventually get rid of VideoWorldApp class, and build the routing map directly in the servlet.  This would avoid passing all the repositories to the app class.
One unfortunate consequence is that if the logoutAction is outside of the app, there is no way anymore to test that it is protected.